Remove extra layer of abstraction
Having an extra layer between the store and AllStoreHolder caused a
significant regression probably due to the call sites being megamorhphic
preventing inlining to take place.
