+++ community/kernel/src/main/java/org/neo4j/kernel/impl/storageengine/impl/recordstorage/StoreStatement.java
@@ -19,13 +19,10 @@
  */
 package org.neo4j.kernel.impl.storageengine.impl.recordstorage;
 
-import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 import java.util.function.IntPredicate;
 import java.util.function.Supplier;
 
 import org.neo4j.cursor.Cursor;
-import org.neo4j.io.pagecache.PageCursor;
 import org.neo4j.kernel.api.AssertOpen;
 import org.neo4j.kernel.api.exceptions.index.IndexNotFoundKernelException;
 import org.neo4j.kernel.api.schema.index.IndexDescriptor;
@@ -39,26 +36,13 @@ import org.neo4j.kernel.impl.api.store.StoreSinglePropertyCursor;
 import org.neo4j.kernel.impl.api.store.StoreSingleRelationshipCursor;
 import org.neo4j.kernel.impl.locking.Lock;
 import org.neo4j.kernel.impl.locking.LockService;
-import org.neo4j.kernel.impl.store.AbstractDynamicStore;
-import org.neo4j.kernel.impl.store.DynamicArrayStore;
-import org.neo4j.kernel.impl.store.DynamicStringStore;
-import org.neo4j.kernel.impl.store.InvalidRecordException;
 import org.neo4j.kernel.impl.store.NeoStores;
 import org.neo4j.kernel.impl.store.NodeStore;
 import org.neo4j.kernel.impl.store.PropertyStore;
-import org.neo4j.kernel.impl.store.RecordCursor;
 import org.neo4j.kernel.impl.store.RecordCursors;
-import org.neo4j.kernel.impl.store.RecordStore;
 import org.neo4j.kernel.impl.store.RelationshipGroupStore;
 import org.neo4j.kernel.impl.store.RelationshipStore;
 import org.neo4j.kernel.impl.store.StoreType;
-import org.neo4j.kernel.impl.store.record.AbstractBaseRecord;
-import org.neo4j.kernel.impl.store.record.DynamicRecord;
-import org.neo4j.kernel.impl.store.record.NodeRecord;
-import org.neo4j.kernel.impl.store.record.PropertyRecord;
-import org.neo4j.kernel.impl.store.record.RecordLoad;
-import org.neo4j.kernel.impl.store.record.RelationshipGroupRecord;
-import org.neo4j.kernel.impl.store.record.RelationshipRecord;
 import org.neo4j.kernel.impl.util.InstanceCache;
 import org.neo4j.storageengine.api.Direction;
 import org.neo4j.storageengine.api.NodeItem;
@@ -67,9 +51,6 @@ import org.neo4j.storageengine.api.RelationshipItem;
 import org.neo4j.storageengine.api.StorageStatement;
 import org.neo4j.storageengine.api.schema.IndexReader;
 import org.neo4j.storageengine.api.schema.LabelScanReader;
-
-import static org.neo4j.kernel.impl.store.record.AbstractBaseRecord.NO_ID;
-import static org.neo4j.kernel.impl.store.record.RecordLoad.NORMAL;
 
 /**
  * Statement for store layer. This allows for acquisition of cursors on the store data.
@@ -97,19 +78,12 @@ public class StoreStatement implements StorageStatement
     private final RecordCursors recordCursors;
     private final Supplier<LabelScanReader> labelScanStore;
     private final RecordStorageCommandCreationContext commandCreationContext;
-    private final DynamicArrayStore propertyArrayStore;
-    private final DynamicStringStore propertyStringStore;
 
     private IndexReaderFactory indexReaderFactory;
     private LabelScanReader labelScanReader;
 
     private boolean acquired;
     private boolean closed;
-
-    private final Nodes nodes;
-    private final Relationships relationships;
-    private final Groups groups;
-    private final Properties properties;
 
     public StoreStatement( NeoStores neoStores, Supplier<IndexReaderFactory> indexReaderFactory,
             Supplier<LabelScanReader> labelScanReaderSupplier, LockService lockService,
@@ -124,8 +98,6 @@ public class StoreStatement implements StorageStatement
         this.relationshipStore = neoStores.getRelationshipStore();
         this.relationshipGroupStore = neoStores.getRelationshipGroupStore();
         this.propertyStore = neoStores.getPropertyStore();
-        this.propertyArrayStore = propertyStore.getArrayStore();
-        this.propertyStringStore = propertyStore.getStringStore();
         this.recordCursors = new RecordCursors( neoStores );
 
         singleNodeCursor = new InstanceCache<StoreSingleNodeCursor>()
@@ -180,11 +152,6 @@ public class StoreStatement implements StorageStatement
                 return new StorePropertyCursor( recordCursors, this );
             }
         };
-
-        nodes = new Nodes();
-        relationships = new Relationships();
-        groups = new Groups();
-        properties = new Properties();
     }
 
     @Override
@@ -321,188 +288,24 @@ public class StoreStatement implements StorageStatement
     @Override
     public Nodes nodes()
     {
-        return nodes;
+        return nodeStore;
     }
 
     @Override
     public Relationships relationships()
     {
-        return relationships;
+        return relationshipStore;
     }
 
     @Override
     public Groups groups()
     {
-        return groups;
+        return relationshipGroupStore;
     }
 
     @Override
     public Properties properties()
     {
-        return properties;
-    }
-
-    class Nodes implements StorageStatement.Nodes
-    {
-        @Override
-        public PageCursor openPageCursor( long reference )
-        {
-            return nodeStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public void loadRecordByCursor( long reference, NodeRecord nodeRecord, RecordLoad mode, PageCursor cursor )
-                throws InvalidRecordException
-        {
-            nodeStore.getRecordByCursor( reference, nodeRecord, mode, cursor );
-        }
-
-        @Override
-        public long getHighestPossibleIdInUse()
-        {
-            return nodeStore.getHighestPossibleIdInUse();
-        }
-
-        @Override
-        public RecordCursor<DynamicRecord> newLabelCursor()
-        {
-            return newCursor( nodeStore.getDynamicLabelStore() );
-        }
-    }
-
-    class Relationships implements StorageStatement.Relationships
-    {
-        @Override
-        public PageCursor openPageCursor( long reference )
-        {
-            return relationshipStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public void loadRecordByCursor( long reference, RelationshipRecord relationshipRecord, RecordLoad mode,
-                PageCursor cursor ) throws InvalidRecordException
-        {
-            relationshipStore.getRecordByCursor( reference, relationshipRecord, mode, cursor );
-        }
-
-        @Override
-        public long getHighestPossibleIdInUse()
-        {
-            return relationshipStore.getHighestPossibleIdInUse();
-        }
-    }
-
-    class Groups implements StorageStatement.Groups
-    {
-        @Override
-        public PageCursor openPageCursor( long reference )
-        {
-            return relationshipGroupStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public void loadRecordByCursor( long reference, RelationshipGroupRecord relationshipGroupRecord,
-                RecordLoad mode, PageCursor cursor ) throws InvalidRecordException
-        {
-            relationshipGroupStore.getRecordByCursor( reference, relationshipGroupRecord, mode, cursor );
-        }
-
-        @Override
-        public long getHighestPossibleIdInUse()
-        {
-            return relationshipGroupStore.getHighestPossibleIdInUse();
-        }
-    }
-
-    class Properties implements StorageStatement.Properties
-    {
-        @Override
-        public PageCursor openPageCursor( long reference )
-        {
-            return propertyStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public void loadRecordByCursor( long reference, PropertyRecord propertyBlocks, RecordLoad mode,
-                PageCursor cursor ) throws InvalidRecordException
-        {
-            propertyStore.getRecordByCursor( reference, propertyBlocks, mode, cursor );
-        }
-
-        @Override
-        public long getHighestPossibleIdInUse()
-        {
-            return propertyStore.getHighestPossibleIdInUse();
-        }
-
-        @Override
-        public PageCursor openStringPageCursor( long reference )
-        {
-            return propertyStringStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public PageCursor openArrayPageCursor( long reference )
-        {
-            return propertyArrayStore.openPageCursorForReading( reference );
-        }
-
-        @Override
-        public ByteBuffer loadString( long reference, ByteBuffer buffer, PageCursor page )
-        {
-            return readDynamic( propertyStore.getStringStore(), reference, buffer, page );
-        }
-
-        @Override
-        public ByteBuffer loadArray( long reference, ByteBuffer buffer, PageCursor page )
-        {
-            return readDynamic( propertyStore.getArrayStore(), reference, buffer, page );
-        }
-    }
-
-    private static ByteBuffer readDynamic( AbstractDynamicStore store, long reference, ByteBuffer buffer,
-            PageCursor page )
-    {
-        if ( buffer == null )
-        {
-            buffer = ByteBuffer.allocate( 512 );
-        }
-        else
-        {
-            buffer.clear();
-        }
-        DynamicRecord record = store.newRecord();
-        do
-        {
-            //We need to load forcefully here since otherwise we can have inconsistent reads
-            //for properties across blocks, see org.neo4j.graphdb.ConsistentPropertyReadsIT
-            store.getRecordByCursor( reference, record, RecordLoad.FORCE, page );
-            reference = record.getNextBlock();
-            byte[] data = record.getData();
-            if ( buffer.remaining() < data.length )
-            {
-                buffer = grow( buffer, data.length );
-            }
-            buffer.put( data, 0, data.length );
-        }
-        while ( reference != NO_ID );
-        return buffer;
-    }
-
-    private static ByteBuffer grow( ByteBuffer buffer, int required )
-    {
-        buffer.flip();
-        int capacity = buffer.capacity();
-        do
-        {
-            capacity *= 2;
-        }
-        while ( capacity - buffer.limit() < required );
-        return ByteBuffer.allocate( capacity ).order( ByteOrder.LITTLE_ENDIAN ).put( buffer );
-    }
-
-    private static <R extends AbstractBaseRecord> RecordCursor<R> newCursor( RecordStore<R> store )
-    {
-        return store.newRecordCursor( store.newRecord() ).acquire( store.getNumberOfReservedLowIds(), NORMAL );
+        return propertyStore;
     }
 }
