+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/NodeStore.java
@@ -26,6 +26,7 @@ import java.util.List;
 
 import org.neo4j.io.pagecache.PageCache;
 import org.neo4j.kernel.configuration.Config;
+import org.neo4j.kernel.impl.storageengine.impl.recordstorage.StoreStatement;
 import org.neo4j.kernel.impl.store.format.RecordFormats;
 import org.neo4j.kernel.impl.store.id.IdGeneratorFactory;
 import org.neo4j.kernel.impl.store.id.IdType;
@@ -40,7 +41,7 @@ import static org.neo4j.kernel.impl.store.NoStoreHeaderFormat.NO_STORE_HEADER_FO
 /**
  * Implementation of the node store.
  */
-public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader>
+public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader> implements StoreStatement.Nodes
 {
     public static Long readOwnerFromDynamicLabelsRecord( DynamicRecord record )
     {
@@ -55,6 +56,13 @@ public class NodeStore extends CommonAbstractStore<NodeRecord,NoStoreHeader>
         }
         Bits bits = Bits.bitsFromBytes( array );
         return bits.getLong( requiredBits );
+    }
+
+    @Override
+    public RecordCursor<DynamicRecord> newLabelCursor()
+    {
+        return dynamicLabelStore.newRecordCursor( dynamicLabelStore.newRecord() ).acquire( getNumberOfReservedLowIds(),
+                RecordLoad.NORMAL );
     }
 
     public abstract static class Configuration
